+++ app/searchers/quick_search/world_cat_discovery_api_searcher.rb
@@ -68,4 +68,26 @@ module QuickSearch
       allowed_values.last
     end
   end
+
+  # Provides a mapping of WorldCat Bibliographic book formats
+  # to simple and more normalized types
+  class ItemFormats
+    # Map of Bibligraphic book_formats to item format
+    @item_formats = {
+      'http://schema.org/Hardcover' => 'book',
+      'http://bibliograph.net/LargePrintBook' => 'book',
+      'http://schema.org/Paperback' => 'book',
+      'http://bibliograph.net/PrintBook' => 'book',
+      'http://bibliograph.net/AudioBook' => 'audio_book',
+      'http://schema.org/EBook' => 'e_book'
+    }
+
+    # Returns string representing the item format for the given
+    # Bibliographic record
+    def self.item_format(bib)
+      book_format = bib.book_format.to_s
+
+      @item_formats[book_format] || 'other'
+    end
+  end
 end
